Refactor as gem (preliminary)
