+++ Gemfile
@@ -1,11 +1,5 @@
+# frozen_string_literal: true
+
 source "https://rubygems.org/"
 
-gem "nokogiri"
-gem "mechanize"
-gem "reverse_adoc"
-gem "vcr"
-gem "webmock"
-gem "rspec"
-gem "rake"
-
-gem "pry"
+gemspec
