--- README.md
@@ -1,5 +1,0 @@
-USAGE
-=====
-
-    $ bundle install
-    $ ruby lib/bipm.rb
