+++ lib/bipm-data-importer.rb
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+require_relative "bipm/data/importer"
