+++ lib/bipm/data/importer.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+require_relative "importer/asciimath"
+require_relative "importer/common"
+require_relative "importer/version"
+require "date"
+
+module Bipm
+  module Data
+    module Importer
+      class Error < StandardError; end
+
+    end
+  end
+end
