+++ lib/bipm/data/importer/version.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module Bipm
+  module Data
+    module Importer
+      VERSION = "0.1.0"
+    end
+  end
+end
