+++ sig/bipm/data/importer.rbs
@@ -1,0 +1,8 @@
+module Bipm
+  module Data
+    module Importer
+      VERSION: String
+      # See the writing guide of rbs: https://github.com/ruby/rbs#guides
+    end
+  end
+end
