+++ spec/bipm/data/importer_spec.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+RSpec.describe Bipm::Data::Importer do
+  it "has a version number" do
+    expect(Bipm::Data::Importer::VERSION).not_to be nil
+  end
+
+end
