+++ spec/spec_helper.rb
@@ -1,11 +1,10 @@
-require 'bundler/setup'
-require 'rspec/matchers'
-require 'yaml'
-require 'pry'
+# frozen_string_literal: true
+
+require_relative "../lib/bipm-data-importer"
 
 RSpec.configure do |config|
   # Enable flags like --only-failures and --next-failure
-  config.example_status_persistence_file_path = '.rspec_status'
+  config.example_status_persistence_file_path = ".rspec_status"
 
   # Disable RSpec exposing methods globally on `Module` and `main`
   config.disable_monkey_patching!
